requirements.txt
appdirs==1.4.3
click==6.7
dj-database-url==0.4.2
Django==1.11
django-appconf==1.0.2
django-classy-tags==0.8.0
django-twitter-tag==1.2.1
djangorestframework==3.6.3
gunicorn==19.7.1
packaging==16.8
pkg-resources==0.0.0
psycopg2==2.7.1
python-dotenv==0.6.4
pytz==2017.2
six==1.10.0
twitter==1.17.1
whitenoise==3.3.0
